requirements.txt
backcall==0.1.0
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.9
django-heroku==0.3.1
gunicorn==19.9.0
jedi==0.13.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
psycopg2==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.7
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.6
wcwidth==0.1.7
whitenoise==4.1
factory-boy==2.11.1
